+++ src/main/java/de/maxhenkel/easyvillagers/items/VillagerItem.java
@@ -98,7 +98,7 @@ public class VillagerItem extends Item {
         }
         if (world.rand.nextInt(20) == 0) {
             PlayerEntity playerEntity = (PlayerEntity) entity;
-            playerEntity.playSound(SoundEvents.ENTITY_VILLAGER_AMBIENT, SoundCategory.NEUTRAL, 1F, 1F);
+            playerEntity.playSound(SoundEvents.ENTITY_VILLAGER_AMBIENT, SoundCategory.BLOCKS, 1F, 1F);
         }
     }